remove useless function in wilcox wall function. BC for omega is now imposed at the wall boundary and not at the first cell center
