--- wallFunctions/wilcoxOmegaWallFunction/WilcoxomegaWallFunctionFvPatchScalarField.C
@@ -1,606 +1,0 @@
-/*---------------------------------------------------------------------------*\
-  =========                 |
-  \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox
-   \\    /   O peration     |
-    \\  /    A nd           | Copyright (C) 2011-2016 OpenFOAM Foundation
-     \\/     M anipulation  |
--------------------------------------------------------------------------------
-License
-    This file is part of OpenFOAM.
-
-    OpenFOAM is free software: you can redistribute it and/or modify it
-    under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    OpenFOAM is distributed in the hope that it will be useful, but WITHOUT
-    ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-    FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
-    for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with OpenFOAM.  If not, see <http://www.gnu.org/licenses/>.
-
-\*---------------------------------------------------------------------------*/
-
-#include "WilcoxomegaWallFunctionFvPatchScalarField.H"
-#include "turbulenceModel.H"
-#include "fvPatchFieldMapper.H"
-#include "fvMatrix.H"
-#include "volFields.H"
-#include "wallFvPatch.H"
-#include "nutkWallFunctionFvPatchScalarField.H"
-#include "addToRunTimeSelectionTable.H"
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace Foam
-{
-
-// * * * * * * * * * * * * * * Static Data Members * * * * * * * * * * * * * //
-
-scalar WilcoxomegaWallFunctionFvPatchScalarField::tolerance_ = 1e-5;
-
-// * * * * * * * * * * * * Protected Member Functions  * * * * * * * * * * * //
-
-void WilcoxomegaWallFunctionFvPatchScalarField::checkType()
-{
-    if (not isA<wallFvPatch>(patch()))
-    {
-        FatalErrorInFunction
-            << "Invalid wall function specification" << nl
-            << "    Patch type for patch " << patch().name()
-            << " must be wall" << nl
-            << "    Current patch type is " << patch().type() << nl << endl
-            << abort(FatalError);
-    }
-}
-
-void WilcoxomegaWallFunctionFvPatchScalarField::writeLocalEntries(Ostream& os)
-const
-{
-    os.writeKeyword("kn") << kn_ << token::END_STATEMENT << nl;
-//    os.writeKeyword("kappa") << kappa_ << token::END_STATEMENT << nl;
-//    os.writeKeyword("E") << E_ << token::END_STATEMENT << nl;
-//    os.writeKeyword("beta1") << beta1_ << token::END_STATEMENT << nl;
-}
-
-
-void WilcoxomegaWallFunctionFvPatchScalarField::setMaster()
-{
-    if (master_ != -1)
-    {
-        return;
-    }
-
-    const volScalarField& omega =
-        static_cast<const volScalarField&>(this->internalField());
-
-    const volScalarField::Boundary& bf = omega.boundaryField();
-
-    label master = -1;
-    forAll(bf, patchi)
-    {
-        if (isA<WilcoxomegaWallFunctionFvPatchScalarField>(bf[patchi]))
-        {
-            WilcoxomegaWallFunctionFvPatchScalarField& opf = omegaPatch(patchi);
-
-            if (master == -1)
-            {
-                master = patchi;
-            }
-
-            opf.master() = master;
-        }
-    }
-}
-
-
-void WilcoxomegaWallFunctionFvPatchScalarField::createAveragingWeights()
-{
-    const volScalarField& omega =
-        static_cast<const volScalarField&>(this->internalField());
-
-    const volScalarField::Boundary& bf = omega.boundaryField();
-
-    const fvMesh& mesh = omega.mesh();
-
-    if (initialised_ && (mesh.changing() == false))
-    {
-        return;
-    }
-
-    volScalarField weights
-    (
-        IOobject
-        (
-            "weights",
-            mesh.time().timeName(),
-            mesh,
-            IOobject::NO_READ,
-            IOobject::NO_WRITE,
-            false // do not register
-        ),
-        mesh,
-        dimensionedScalar("zero", dimless, 0.0)
-    );
-
-    DynamicList<label> omegaPatches(bf.size());
-    forAll(bf, patchi)
-    {
-        if (isA<WilcoxomegaWallFunctionFvPatchScalarField>(bf[patchi]))
-        {
-            omegaPatches.append(patchi);
-
-            const labelUList& faceCells = bf[patchi].patch().faceCells();
-            forAll(faceCells, i)
-            {
-                label celli = faceCells[i];
-                weights[celli]++;
-            }
-        }
-    }
-
-    cornerWeights_.setSize(bf.size());
-    forAll(omegaPatches, i)
-    {
-        label patchi = omegaPatches[i];
-        const fvPatchScalarField& wf = weights.boundaryField()[patchi];
-        cornerWeights_[patchi] = 1.0/wf.patchInternalField();
-    }
-
-    G_.setSize(internalField().size(), 0.0);
-    omega_.setSize(internalField().size(), 0.0);
-
-    initialised_ = true;
-}
-
-
-WilcoxomegaWallFunctionFvPatchScalarField&
-WilcoxomegaWallFunctionFvPatchScalarField::omegaPatch(const label patchi)
-{
-    const volScalarField& omega =
-        static_cast<const volScalarField&>(this->internalField());
-
-    const volScalarField::Boundary& bf = omega.boundaryField();
-
-    const WilcoxomegaWallFunctionFvPatchScalarField& opf =
-        refCast<const WilcoxomegaWallFunctionFvPatchScalarField>(bf[patchi]);
-
-    return const_cast<WilcoxomegaWallFunctionFvPatchScalarField&>(opf);
-}
-
-
-void WilcoxomegaWallFunctionFvPatchScalarField::calculateTurbulenceFields
-(
-    const turbulenceModel& turbModel,
-    scalarField& G0,
-    scalarField& omega0
-)
-{
-    // accumulate all of the G and omega contributions
-    forAll(cornerWeights_, patchi)
-    {
-        if (cornerWeights_[patchi].empty() == false)
-        {
-            WilcoxomegaWallFunctionFvPatchScalarField& opf = omegaPatch(patchi);
-
-            const List<scalar>& w = cornerWeights_[patchi];
-
-            opf.calculate(turbModel, w, opf.patch(), G0, omega0);
-        }
-    }
-
-    // apply zero-gradient condition for omega
-    forAll(cornerWeights_, patchi)
-    {
-        if (cornerWeights_[patchi].empty() == false)
-        {
-            WilcoxomegaWallFunctionFvPatchScalarField& opf = omegaPatch(patchi);
-
-            opf == scalarField(omega0, opf.patch().faceCells());
-        }
-    }
-}
-
-
-void WilcoxomegaWallFunctionFvPatchScalarField::calculate
-(
-    const turbulenceModel& turbModel,
-    const List<scalar>& cornerWeights,
-    const fvPatch& patch,
-    scalarField& G,
-    scalarField& omega
-)
-{
-    const label patchi = patch.index();
-
-    const tmp<volScalarField> tk = turbModel.k();
-
-    const tmp<scalarField> tnuw = turbModel.nu(patchi);
-    const scalarField& nuw = tnuw();
-
-    const tmp<scalarField> tnutw = turbModel.nut(patchi);
-    const tmp<volScalarField> tnut = turbModel.nut();
-    const scalarField& nutw = tnutw();
-
-    const fvPatchVectorField& Uw = turbModel.U().boundaryField()[patchi];
-
-    const scalarField magGradUw(mag(Uw.snGrad()));
-
-    // Set omega and G
-    forAll(nutw, facei)
-    {
-        label celli = patch.faceCells()[facei];
-
-        scalar w = cornerWeights[facei];
-
-/*
-        scalar omegaVis = 6.0*nuw[facei]/(beta1_*sqr(y[facei]));
-
-        scalar omegaLog = sqrt(k[celli])/(Cmu25*kappa_*y[facei]);
-
-        omega[celli] += w*sqrt(sqr(omegaVis) + sqr(omegaLog));
-
-        G[celli] +=
-            w
-           *(nutw[facei] + nuw[facei])
-           *magGradUw[facei]
-           *Cmu25*sqrt(k[celli])
-           /(kappa_*y[facei]);
-*/
-        scalar utau = max(1.0e-6, sqrt((nutw[facei] + nuw[facei])*
-                      magGradUw[facei]));
-        scalar knplus = kn_*utau/nuw[facei];
-        scalar SR = 0e0;
-        //Info<<utau<<endl;
-        //Info<<knplus<<endl;
-        if (knplus<=25e0)
-          {
-            SR = pow(50e0/knplus, 2);
-          }
-        else
-          {
-            SR = 100e0/knplus;
-          }
-
-        omega[celli] +=  w*pow(utau, 2)*SR/nuw[facei];
-    }
-}
-
-
-// * * * * * * * * * * * * * * * * Constructors  * * * * * * * * * * * * * * //
-
-WilcoxomegaWallFunctionFvPatchScalarField::
-WilcoxomegaWallFunctionFvPatchScalarField
-(
-    const fvPatch& p,
-    const DimensionedField<scalar, volMesh>& iF
-)
-:
-    fixedValueFvPatchField<scalar>(p, iF),
-    kn_(1e-6),
-    G_(),
-    omega_(),
-    initialised_(false),
-    master_(-1),
-    cornerWeights_()
-{
-    checkType();
-}
-
-
-WilcoxomegaWallFunctionFvPatchScalarField::
-WilcoxomegaWallFunctionFvPatchScalarField
-(
-    const WilcoxomegaWallFunctionFvPatchScalarField& ptf,
-    const fvPatch& p,
-    const DimensionedField<scalar, volMesh>& iF,
-    const fvPatchFieldMapper& mapper
-)
-:
-    fixedValueFvPatchField<scalar>(ptf, p, iF, mapper),
-    kn_(ptf.kn_),
-    G_(),
-    omega_(),
-    initialised_(false),
-    master_(-1),
-    cornerWeights_()
-{
-    checkType();
-}
-
-
-WilcoxomegaWallFunctionFvPatchScalarField::
-WilcoxomegaWallFunctionFvPatchScalarField
-(
-    const fvPatch& p,
-    const DimensionedField<scalar, volMesh>& iF,
-    const dictionary& dict
-)
-:
-    fixedValueFvPatchField<scalar>(p, iF, dict),
-    kn_(dict.lookupOrDefault<scalar>("kn", 1e-6)),
-    G_(),
-    omega_(),
-    initialised_(false),
-    master_(-1),
-    cornerWeights_()
-{
-    checkType();
-
-    // apply zero-gradient condition on start-up
-    this->operator==(patchInternalField());
-}
-
-
-WilcoxomegaWallFunctionFvPatchScalarField::
-WilcoxomegaWallFunctionFvPatchScalarField
-(
-    const WilcoxomegaWallFunctionFvPatchScalarField& owfpsf
-)
-:
-    fixedValueFvPatchField<scalar>(owfpsf),
-    kn_(owfpsf.kn_),
-    G_(),
-    omega_(),
-    initialised_(false),
-    master_(-1),
-    cornerWeights_()
-{
-    checkType();
-}
-
-
-WilcoxomegaWallFunctionFvPatchScalarField::
-WilcoxomegaWallFunctionFvPatchScalarField
-(
-    const WilcoxomegaWallFunctionFvPatchScalarField& owfpsf,
-    const DimensionedField<scalar, volMesh>& iF
-)
-:
-    fixedValueFvPatchField<scalar>(owfpsf, iF),
-    kn_(owfpsf.kn_),
-    G_(),
-    omega_(),
-    initialised_(false),
-    master_(-1),
-    cornerWeights_()
-{
-    checkType();
-}
-
-
-// * * * * * * * * * * * * * * * Member Functions  * * * * * * * * * * * * * //
-
-scalarField& WilcoxomegaWallFunctionFvPatchScalarField::G(bool init)
-{
-    if (patch().index() == master_)
-    {
-        if (init)
-        {
-            G_ = 0.0;
-        }
-
-        return G_;
-    }
-
-    return omegaPatch(master_).G();
-}
-
-
-scalarField& WilcoxomegaWallFunctionFvPatchScalarField::omega(bool init)
-{
-    if (patch().index() == master_)
-    {
-        if (init)
-        {
-            omega_ = 0.0;
-        }
-
-        return omega_;
-    }
-
-    return omegaPatch(master_).omega(init);
-}
-
-
-void WilcoxomegaWallFunctionFvPatchScalarField::updateCoeffs()
-{
-    if (updated())
-    {
-        return;
-    }
-
-    const turbulenceModel& turbModel = db().lookupObject<turbulenceModel>
-    (
-        IOobject::groupName
-        (
-            turbulenceModel::propertiesName,
-            internalField().group()
-        )
-    );
-
-    setMaster();
-
-    if (patch().index() == master_)
-    {
-        createAveragingWeights();
-        calculateTurbulenceFields(turbModel, G(true), omega(true));
-    }
-
-    const scalarField& G0 = this->G();
-    const scalarField& omega0 = this->omega();
-
-    typedef DimensionedField<scalar, volMesh> FieldType;
-
-    FieldType& G =
-        const_cast<FieldType&>
-        (
-            db().lookupObject<FieldType>(turbModel.GName())
-        );
-
-    FieldType& omega = const_cast<FieldType&>(internalField());
-
-    forAll(*this, facei)
-    {
-        label celli = patch().faceCells()[facei];
-
-        G[celli] = G0[celli];
-        omega[celli] = omega0[celli];
-    }
-
-    fvPatchField<scalar>::updateCoeffs();
-}
-
-
-void WilcoxomegaWallFunctionFvPatchScalarField::updateWeightedCoeffs
-(
-    const scalarField& weights
-)
-{
-    if (updated())
-    {
-        return;
-    }
-
-    const turbulenceModel& turbModel = db().lookupObject<turbulenceModel>
-    (
-        IOobject::groupName
-        (
-            turbulenceModel::propertiesName,
-            internalField().group()
-        )
-    );
-
-    setMaster();
-
-    if (patch().index() == master_)
-    {
-        createAveragingWeights();
-        calculateTurbulenceFields(turbModel, G(true), omega(true));
-    }
-
-    const scalarField& G0 = this->G();
-    const scalarField& omega0 = this->omega();
-
-    typedef DimensionedField<scalar, volMesh> FieldType;
-
-    FieldType& G =
-        const_cast<FieldType&>
-        (
-            db().lookupObject<FieldType>(turbModel.GName())
-        );
-
-    FieldType& omega = const_cast<FieldType&>(internalField());
-
-    scalarField& omegaf = *this;
-
-    // only set the values if the weights are > tolerance
-    forAll(weights, facei)
-    {
-        scalar w = weights[facei];
-
-        if (w > tolerance_)
-        {
-            label celli = patch().faceCells()[facei];
-
-            G[celli] = (1.0 - w)*G[celli] + w*G0[celli];
-            omega[celli] = (1.0 - w)*omega[celli] + w*omega0[celli];
-            omegaf[facei] = omega[celli];
-        }
-    }
-
-    fvPatchField<scalar>::updateCoeffs();
-}
-
-
-void WilcoxomegaWallFunctionFvPatchScalarField::manipulateMatrix
-(
-    fvMatrix<scalar>& matrix
-)
-{
-    if (manipulatedMatrix())
-    {
-        return;
-    }
-
-    matrix.setValues(patch().faceCells(), patchInternalField());
-
-    fvPatchField<scalar>::manipulateMatrix(matrix);
-}
-
-
-void WilcoxomegaWallFunctionFvPatchScalarField::manipulateMatrix
-(
-    fvMatrix<scalar>& matrix,
-    const Field<scalar>& weights
-)
-{
-    if (manipulatedMatrix())
-    {
-        return;
-    }
-
-    DynamicList<label> constraintCells(weights.size());
-    DynamicList<scalar> constraintomega(weights.size());
-    const labelUList& faceCells = patch().faceCells();
-
-    const DimensionedField<scalar, volMesh>& omega
-        = internalField();
-
-    label nConstrainedCells = 0;
-
-
-    forAll(weights, facei)
-    {
-        // only set the values if the weights are > tolerance
-        if (weights[facei] > tolerance_)
-        {
-            nConstrainedCells++;
-
-            label celli = faceCells[facei];
-
-            constraintCells.append(celli);
-            constraintomega.append(omega[celli]);
-        }
-    }
-
-    if (debug)
-    {
-        Pout<< "Patch: " << patch().name()
-            << ": number of constrained cells = " << nConstrainedCells
-            << " out of " << patch().size()
-            << endl;
-    }
-
-    matrix.setValues
-    (
-        constraintCells,
-        scalarField(constraintomega)
-    );
-
-    fvPatchField<scalar>::manipulateMatrix(matrix);
-}
-
-
-void WilcoxomegaWallFunctionFvPatchScalarField::write(Ostream& os) const
-{
-    writeLocalEntries(os);
-    fixedValueFvPatchField<scalar>::write(os);
-}
-
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-makePatchTypeField
-(
-    fvPatchScalarField,
-    WilcoxomegaWallFunctionFvPatchScalarField
-);
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-} // End namespace Foam
-
-// ************************************************************************* //
